requirements.txt
flask==2.3.3
scrapy==2.11.2
pony==0.7.17
PyMySQL==1.1.1
py-pretty==1
pycryptodome==3.20.0
elasticsearch>=7.0.0,<8.0.0
elasticsearch-dsl>=7.0.0,<8.0.0
python-dateutil==2.8.2
Twisted==22.10.0
PySocks==1.7.1
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.2
tabulate==0.9.0
langdetect==1.0.9
pycountry==23.12.11
networkx==3.2.1
python-memcached==1.62
pillow==10.3.0
paramiko==3.4.0
werkzeug==2.3.7
timeout-decorator==0.5.0
bitcoin==1.1.42
markupsafe==2.1.5
